feat: Refactored code of organizations
* Moved code realted to organizations from register.go and
  register_test.go to organization.go and organization_test.go
* Added one more unit tests for the case server returns
  corrupted data
